Fix lints
