+++ ci/pipeline/ray_release_test.py
@@ -1,4 +1,5 @@
 import boto3
+import json
 import os
 import subprocess
 import sys
@@ -11,7 +12,7 @@ from ray_coverage import (
     S3_BUCKET_DIR,
     S3_BUCKET_NAME,
 )
-from typing import List
+from typing import List, Set
 
 
 @click.command()
@@ -34,21 +35,43 @@ def main(artifact_dir: str) -> int:
     logger.info(f"Changed files: {changed_files}")
     test_targets = _get_test_targets_for_changed_files(changed_files, artifact_dir)
     logger.info(test_targets)
+    _run_tests(test_targets)
     return 0
 
 
-def _get_test_targets_for_changed_files(
-    changed_files: List[str],
-    artifact_dir: str,
-) -> str:
+def _run_tests(test_targets: Set[str]) -> None:
+    """
+    Run the tests.
+    """
+    subprocess.check_call(
+        ["bazel", "test", "--test_output=streamed"] + list(test_targets)
+    )
+
+
+def _get_test_targets_for_changed_files(changed_files: List[str]) -> str:
     """
     Get the test target for the changed files.
     """
-    coverage_file = _get_coverage_file(artifact_dir)
-    coverage_info = subprocess.check_output(
-        ["coverage", "report", f"--data-file={coverage_file}"]
-    ).decode("utf-8")
-    return coverage_info
+    coverage_file = _get_coverage_file()
+    subprocess.check_output(
+        [
+            "coverage",
+            "json",
+            f"--data-file={coverage_file}",
+            "--show-contexts",
+            f"--include={','.join(changed_files)}",
+        ]
+    )
+    coverage_data = json.load(open("coverage.json"))
+    test_targets = set()
+    for file_metadata in coverage_data["files"].values():
+        contexts = file_metadata["contexts"]
+        for tests_per_line in contexts.values():
+            for test_per_line in tests_per_line:
+                test_targets.add(
+                    f"//release:{test_per_line.split('::')[0].split('/')[-1][:-3]}",
+                )
+    return test_targets
 
 
 def _get_coverage_file(artifact_dir: str) -> str:
